requirements.txt
fastapi
uvicorn
tensorflow
pillow
streamlit
requests
scikit-learn
